requirements.txt
# =============================================================================
# BXMA Risk/Quant Platform - Python Dependencies
# =============================================================================

# Core Data Processing
polars>=0.20.0
pyarrow>=14.0.0
duckdb>=0.9.0
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
torch>=2.1.0
xgboost>=2.0.0
lightgbm>=4.1.0

# Optimization
cvxpy>=1.4.0
cvxopt>=1.3.0
riskfolio-lib>=5.0.0

# Statistical Analysis
statsmodels>=0.14.0
arch>=6.2.0
hmmlearn>=0.3.0

# Explainability
shap>=0.44.0
lime>=0.2.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Caching & Messaging
redis>=5.0.0
aiokafka>=0.10.0
celery>=5.3.0

# Monitoring & Logging
loguru>=0.7.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0

# Visualization (server-side)
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
factory-boy>=3.3.0
faker>=20.0.0

# Development
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
ruff>=0.1.6

# Market Data
yfinance>=0.2.36
aiohttp>=3.9.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
tenacity>=8.2.0
orjson>=3.9.0
